refactor(navigation): restructure AppNavigator and migrate navigation logic to repository
- Remove AppNavigator interface and relocate implementation to feat-navigation's internal data package
- Update NavigationRepository to delegate navigation actions to internal AppNavigator instance
- Change NavHost composable to bind navigation state via NavigationRepository instead of AppNavigator directly
- Clean up imports and references to follow new architecture with encapsulated navigation controller
- Remove obsolete activity_ic.xml and other unused vector drawable resources from feat-navigation
- Move shared navigation icons to app module's drawable resources
- Simplify NavigationData by removing unused account type field
- Update Koin module to provide concrete AppNavigator and bind it internally within NavigationRepository
- Revise documentation and sample usage to reflect internal-only AppNavigator usage and repository as public API
- Add missing string resources to app module and remove duplicates from feat-navigation resources
