+++ app/src/main/java/com/skewnexus/trackhub/navigation/components/BottomNavigationBar.kt
@@ -21,8 +21,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import com.greenvenom.core_navigation.routes.Destination
 import com.greenvenom.core_navigation.utils.NavigationType
+import com.skewnexus.trackhub.R
 import com.trackhub.feat_hub.presentation.routes.HubDest
-import com.trackhub.feat_navigation.R
 
 @Composable
 fun BottomNavigationBar(
