+++ navigation/core-navigation/src/main/java/com/greenvenom/core_navigation/domain/repos/NavigationData.kt
@@ -5,7 +5,7 @@ import com.greenvenom.core_navigation.routes.Destination
 
 /**
  * Data class representing the state of navigation UI.
- * Holds current and previous destinations, flags for side/bottom/top bar visibility, and account type index.
+ * Holds current and previous destinations and flags for side/bottom/top bar visibility.
  */
 @Immutable
 data class NavigationData(
@@ -13,6 +13,5 @@ data class NavigationData(
     val previousDestination: Destination? = null,
     val isCurrentDestinationSide: Boolean = false,
     val bottomBarState: Boolean = false,
-    val topBarState: Boolean = false,
-    val accountTypeIndex: Int = 0
+    val topBarState: Boolean = false
 )
