+++ navigation/core-navigation/src/main/java/com/greenvenom/core_navigation/utils/NavigationType.kt
@@ -4,7 +4,7 @@ import com.greenvenom.core_navigation.routes.Destination
 
 /**
  * Represents a navigation action that can be performed.
- * Used by the [NavigationRepository] to trigger navigation via the [AppNavigator].
+ * Used by the [NavigationRepository] to trigger navigation via the [com.trackhub.feat_navigation.data.AppNavigator].
  */
 sealed class NavigationType {
     /** Navigate back to the previous destination in the current back stack. */
