+++ navigation/feat-navigation/src/main/java/com/trackhub/feat_navigation/data/repos/NavigationRepositoryImpl.kt
@@ -1,10 +1,11 @@
 package com.trackhub.feat_navigation.data.repos
 
+import com.greenvenom.core_navigation.domain.NavigationState
 import com.greenvenom.core_navigation.domain.repos.NavigationData
 import com.greenvenom.core_navigation.domain.repos.NavigationRepository
 import com.greenvenom.core_navigation.routes.DestinationType
-import com.greenvenom.core_navigation.utils.AppNavigator
 import com.greenvenom.core_navigation.utils.NavigationType
+import com.trackhub.feat_navigation.data.AppNavigator
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
@@ -14,6 +15,11 @@ class NavigationRepositoryImpl(
 ): NavigationRepository {
     private val _navigationData = MutableStateFlow(NavigationData())
     override val navigationData = _navigationData.asStateFlow()
+
+    override fun bind(navigationState: NavigationState) {
+        appNavigator.bind(navigationState)
+        updateStoredDestinations()
+    }
 
     override fun navigate(navigationType: NavigationType) {
         when(navigationType) {
