+++ navigation/feat-navigation/src/main/res/values-ar/strings.xml
@@ -1,10 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="my_hubs">مخازني</string>
-    <string name="notifications">الإشعارات</string>
-    <string name="profile">الملف الشخصي</string>
-    <string name="shared_hubs">مخازن الآخرين</string>
-    <string name="menu">المزيد</string>
-    <string name="back_button">الرجوع للخلف</string>
-    <string name="navigation_icon">أيقونة التنقل %1$s</string>
+
 </resources>
